go mad

+++ tests/database/test_dbs.py
@@ -31,8 +31,9 @@ class TestComparability(TestCase):
                 pass
 
     def tearDown(self) -> None:
-        if os.uname().nodename[-3:] == "avd":
-            shutil.rmtree(self.disk_db.path)
+        if os.name == "posix":
+            if os.uname().nodename[-3:] == "avd":
+                shutil.rmtree(self.disk_db.path)
         return super().tearDown()
     
     def test_strings(self):
