+++ tests/database/test_disk.py
@@ -18,8 +18,9 @@ class TestDisk(TestCase):
 
     
     def tearDown(self) -> None:
-        if os.uname().nodename[-3:] == "avd":
-            shutil.rmtree(self.disk_db.path)
+        if os.name == "posix":
+            if os.uname().nodename[-3:] == "avd":
+                shutil.rmtree(self.disk_db.path)
         return super().tearDown()
 
     def test_strings(self):
